requirements.txt
Django
djangorestframework
django-cors-headers
psycopg[c]
celery
openai==1.61.1
google-genai==0.1.0